requirements.txt
requests
Pillow
trimesh
numpy
